Ready for pull request
