+++ .gitignore
@@ -79,4 +79,5 @@ com_crashlytics_export_strings.xml
 crashlytics.properties
 crashlytics-build.properties
 
-
+# Phabulous production config
+config/production.json
