--- phabricator-slack-feed/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "phabricator-slack-feed",
-  "version": "1.1.0",
-  "description": "Phabricator to Slack bridge",
-  "main": "feedProxy.js",
-  "scripts": {
-    "start": "NODE_ENV=production nohup nodejs app > /home/ubuntu/slackbot-$(date --iso-8601=seconds).log 2>&1 &",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Eduardo Trujillo <ed@chromabits.com>",
-  "license": "MIT",
-  "dependencies": {
-    "babel": "^4.7.3",
-    "body-parser": "^1.12.0",
-    "canduit": "^1.0.0",
-    "concat-stream": "^1.4.7",
-    "config": "^1.11.0",
-    "enclosure": "^0.1.3",
-    "express": "^4.11.2",
-    "qs": "^2.3.3",
-    "superagent": "^0.21.0"
-  },
-  "entrypoint": "ProApp",
-  "providers": [
-    "Providers/ConfigServiceProvider",
-    "Providers/DirectoryServiceProvider"
-  ],
-  "autoload": {
-    "roots": [
-      "src"
-    ]
-  }
-}
