--- phabricator-slack-feed/src/Listings.js
@@ -1,8 +1,0 @@
-'use strict';
-
-var Listings =
-{
-    "phabricator-username":    "@slack-username"
-}
-
-module.exports = Listings;
